backend/requirements.txt
fastapi
uvicorn
pydantic
aiohttp
websockets
PyJWT
python-multipart
bcrypt
